+++ src/components/input/validators/min-value.directive.ts
@@ -1,4 +1,4 @@
-import { Directive, Input, forwardRef, OnInit, OnChanges, SimpleChanges } from '@angular/core';
+import { Directive, Input, forwardRef, AfterViewInit, ElementRef } from '@angular/core';
 import { NG_VALIDATORS, Validator, ValidatorFn, AbstractControl, Validators, ValidationErrors } from '@angular/forms';
 
 const MIN_VALIDATOR: any = {
@@ -8,12 +8,18 @@ const MIN_VALIDATOR: any = {
 };
 
 @Directive({
-  selector: 'input[min][type]',
+  selector: 'input[min]',
   providers: [MIN_VALIDATOR]
 })
-export class MinValidatorDirective implements Validator {
+export class MinValidatorDirective implements Validator, AfterViewInit {
   @Input() min: number;
-  @Input() type: string;
+  type: string;
+
+  constructor(private elm: ElementRef) {}
+
+  ngAfterViewInit() {
+    this.type = this.elm.nativeElement.getAttribute('type');
+  }
 
   validate(c: AbstractControl): ValidationErrors | null {
     if (this.type !== 'number') {